% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempConvAllSR.R
\name{F_to_K}
\alias{F_to_K}
\title{Fahrenheit to Kelvin conversion}
\usage{
F_to_K(F_temp2)
}
\arguments{
\item{F_temp2}{The temperature in degrees Fahrenheit}
}
\value{
The temperature in Kelvin
}
\description{
Convert degrees Fahrenheit temperatures to Kelvin
}
\examples{
temp1 <- F_to_K(50);
temp2 <- F_to_K( c(50, 63, 23) );
}
