% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempConvAllSR.R
\name{K_to_C}
\alias{K_to_C}
\title{Kelvin to degrees Celsius conversion}
\usage{
K_to_C(K_temp2)
}
\arguments{
\item{K_temp2}{The temperature in Kelvin}
}
\value{
The temperature in degrees Celsius
}
\description{
Convert Kelvin temperatures to degrees Celsius
}
\examples{
temp1 <- K_to_C(300);
temp2 <- K_to_C( c(280, 312, 323) );
}
