% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempConvAllSR.R
\name{K_to_F}
\alias{K_to_F}
\title{Kelvin to degrees Fahrenheit conversion}
\usage{
K_to_F(K_temp1)
}
\arguments{
\item{K_temp1}{The temperature in Kelvin}
}
\value{
The temperature in degrees Fahrenheit
}
\description{
Convert Kelvin temperatures to degrees Fahrenheit
}
\examples{
temp1 <- K_to_F(300);
temp2 <- K_to_F( c(280, 312, 323) );
}
